refactored out helper functions to utils.rs
